+++ lib/src/dynamics/contacts/chain_and_circle_contact.dart
@@ -1,21 +1,24 @@
 part of forge2d;
 
 class ChainAndCircleContact extends Contact {
-  ChainAndCircleContact(Fixture fA, int indexA, Fixture fB, int indexB)
-      : super(fA, indexA, fB, indexB) {
-    assert(_fixtureA.getType() == ShapeType.CHAIN);
-    assert(_fixtureB.getType() == ShapeType.CIRCLE);
-  }
+  ChainAndCircleContact(
+    Fixture fixtureA,
+    int indexA,
+    Fixture fixtureB,
+    int indexB,
+  )   : assert(fixtureA.getType() == ShapeType.CHAIN),
+        assert(fixtureB.getType() == ShapeType.CIRCLE),
+        super(fixtureA, indexA, fixtureB, indexB);
 
   @override
   void evaluate(Manifold manifold, Transform xfA, Transform xfB) {
-    final ChainShape chain = _fixtureA.getShape() as ChainShape;
-    final EdgeShape edge = chain.getChildEdge(_indexA);
+    final ChainShape chain = fixtureA.shape as ChainShape;
+    final EdgeShape edge = chain.getChildEdge(indexA);
     World.collision.collideEdgeAndCircle(
       manifold,
       edge,
       xfA,
-      _fixtureB.getShape() as CircleShape,
+      fixtureB.shape as CircleShape,
       xfB,
     );
   }