+++ lib/src/dynamics/contacts/chain_and_polygon_contact.dart
@@ -1,21 +1,21 @@
 part of forge2d;
 
 class ChainAndPolygonContact extends Contact {
-  ChainAndPolygonContact(Fixture fA, int indexA, Fixture fB, int indexB)
-      : super(fA, indexA, fB, indexB) {
-    assert(_fixtureA.getType() == ShapeType.CHAIN);
-    assert(_fixtureB.getType() == ShapeType.POLYGON);
-  }
+  ChainAndPolygonContact(
+      Fixture fixtureA, int indexA, Fixture fixtureB, int indexB)
+      : assert(fixtureA.getType() == ShapeType.CHAIN),
+        assert(fixtureB.getType() == ShapeType.POLYGON),
+        super(fixtureA, indexA, fixtureB, indexB);
 
   @override
   void evaluate(Manifold manifold, Transform xfA, Transform xfB) {
-    final ChainShape chain = _fixtureA.getShape() as ChainShape;
-    final EdgeShape edge = chain.getChildEdge(_indexA);
+    final ChainShape chain = fixtureA.shape as ChainShape;
+    final EdgeShape edge = chain.getChildEdge(indexA);
     World.collision.collideEdgeAndPolygon(
       manifold,
       edge,
       xfA,
-      _fixtureB.getShape() as PolygonShape,
+      fixtureB.shape as PolygonShape,
       xfB,
     );
   }